Former-commit-id: 13b654934f94073696a868bac71adb314817ad75 [formerly 61d96543c3a6b6a0e67befe6b76598af52b7fd80] [formerly f52e528f40c92a677b1adaa5652e0225d92d8233 [formerly d64b9ae2b14ff1ddf77d5a8db8acfbde8d109a14]]
Former-commit-id: 4105f1e7cab5302baeb7771155d68ac13bc63729 [formerly af2c730956aa563aade9045ff68ada35166c0880]
Former-commit-id: 88f9a3fadd99bcbee4b77538337659969175e9ad

+++ Frame/SDK/Core/Base/AFTime.hpp
@@ -548,7 +548,7 @@ public:
         AFTime xTime((int)Get(YEAR), (int)Get(MONTH), 1, 0, 0, 0, 0);
         int nFirstDayOfWeek = xTime.GetDayOfWeek();
         int64_t nDaySpan = Get(DAY) - 1;
-        return (nFirstDayOfWeek + nDaySpan) / 7;
+        return int((nFirstDayOfWeek + nDaySpan) / 7);
     }
 
     int GetDayOfYear()
